tests/submission/Infraware/Forms/contents/Forms/input_formtarget.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_formTarget</h3>
	</p>
	
	<hr>

	<div id="log"></div>  

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='hidden' id='input_text'></p>
	</form>
	<script>
	
		var input_text = document.getElementById("input_text");
		input_text.formTarget = "_blank";

		if (typeof(input_text.formTarget) == "string") {
			test(function() {
				assert_equals(input_text.formTarget, "_blank", "formTarget attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("formTarget attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
